+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-16T14:36:32.977Z",
+    "2026-02-16T15:51:45.516Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {